:recycle: Split activitypub service into directory module by responsibility
Pure mechanical move of application/src/service/activitypub.rs into a directory module; no behavior change.
